requirements.txt
aiogram>=3.13,<4.0
asyncpg>=0.29.0
openai>=1.40.0
python-dotenv>=1.0.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
